update: seperate storage and owlbear integration
